+++ src/main.py
@@ -12,16 +12,22 @@ import kivy
 from ethereum.utils import normalize_address
 from kivy.app import App
 from kivy.clock import Clock, mainthread
+from kivy.logger import LOG_LEVELS, Logger
 from kivy.metrics import dp
 from kivy.properties import NumericProperty, ObjectProperty, StringProperty
 from kivy.uix.boxlayout import BoxLayout
 from kivy.uix.floatlayout import FloatLayout
+from kivy.uix.screenmanager import Screen
 from kivy.uix.scrollview import ScrollView
-from kivy.utils import platform
+from kivy.utils import get_color_from_hex, platform
 from kivymd.button import MDFlatButton, MDIconButton
+from kivymd.color_definitions import colors
 from kivymd.dialog import MDDialog
 from kivymd.label import MDLabel
-from kivymd.list import ILeftBodyTouch, OneLineListItem, TwoLineIconListItem
+from kivymd.list import (ILeftBodyTouch, OneLineListItem, TwoLineIconListItem,
+                         TwoLineListItem)
+from kivymd.navigationdrawer import NavigationDrawerHeaderBase
+from kivymd.selectioncontrols import MDSwitch
 from kivymd.snackbar import Snackbar
 from kivymd.textfields import MDTextField
 from kivymd.theming import ThemeManager
@@ -29,9 +35,11 @@ from kivymd.toolbar import Toolbar
 from raven import Client
 from requests.exceptions import ConnectionError
 
-from pywalib import (InsufficientFundsException, NoTransactionFoundException,
-                     PyWalib, UnknownEtherscanException)
+from pywalib import (ROUND_DIGITS, InsufficientFundsException,
+                     NoTransactionFoundException, PyWalib,
+                     UnknownEtherscanException)
 from testsuite import suite
+from version import __version__
 
 kivy.require('1.10.0')
 
@@ -59,6 +67,50 @@ def run_in_thread(fn):
         t.start()
         return t
     return run
+
+
+class NavigationDrawerTwoLineListItem(
+        TwoLineListItem, NavigationDrawerHeaderBase):
+
+    address_property = StringProperty()
+
+    def __init__(self, **kwargs):
+        super(NavigationDrawerTwoLineListItem, self).__init__(**kwargs)
+        Clock.schedule_once(lambda dt: self.setup())
+
+    def setup(self):
+        """
+        Binds Controller.current_account property.
+        """
+        self.controller = App.get_running_app().controller
+        self.controller.bind(
+            current_account=lambda _, value: self.on_current_account(value))
+
+    def on_current_account(self, account):
+        address = "0x" + account.address.encode("hex")
+        self.address_property = address
+
+    def _update_specific_text_color(self, instance, value):
+        pass
+
+    def _set_active(self, active, list):
+        pass
+
+
+class CustomMDSwitch(MDSwitch):
+    """
+    Work around for a MDSwitch bug, refs:
+    https://gitlab.com/kivymd/KivyMD/issues/99
+    """
+
+    def _set_colors(self, *args):
+        """
+        Overrides `MDSwitch._set_colors()` fixes missing attribute
+        `thumb_color_disabled`, refs:
+        https://gitlab.com/kivymd/KivyMD/issues/99
+        """
+        super(CustomMDSwitch, self)._set_colors(*args)
+        self.thumb_color_disabled = get_color_from_hex(colors['Grey']['800'])
 
 
 class IconLeftWidget(ILeftBodyTouch, MDIconButton):
@@ -165,7 +217,7 @@ class Send(BoxLayout):
         controller = App.get_running_app().controller
         pywalib = controller.pywalib
         address = normalize_address(self.send_to_address)
-        amount_eth = self.send_amount
+        amount_eth = round(self.send_amount, ROUND_DIGITS)
         amount_wei = int(amount_eth * pow(10, 18))
         account = controller.pywalib.get_main_account()
         Controller.snackbar_message("Unlocking account...")
@@ -194,42 +246,55 @@ class Send(BoxLayout):
 
 class Receive(BoxLayout):
 
-    current_account = ObjectProperty(None, allownone=True)
-    current_account_string = StringProperty()
+    current_account = ObjectProperty()
+    address_property = StringProperty()
 
     def __init__(self, **kwargs):
         super(Receive, self).__init__(**kwargs)
-        Clock.schedule_once(lambda dt: self.setup())
+        # for some reason setting the timeout to zero
+        # crashes with:
+        # 'super' object has no attribute '__getattr__'
+        # only on first account creation (with auto redirect)
+        # and we cannot yet reproduce in unit tests
+        timeout = 1
+        Clock.schedule_once(lambda dt: self.setup(), timeout)
 
     def setup(self):
         """
-        Default state setup.
+        Binds Controller.current_account property.
         """
         self.controller = App.get_running_app().controller
-        try:
-            self.current_account = self.controller.pywalib.get_main_account()
-        except IndexError:
-            pass
+        self.controller.bind(current_account=self.setter('current_account'))
+        # triggers the update
+        self.current_account = self.controller.current_account
 
     def show_address(self, address):
         self.ids.qr_code_id.data = address
 
-    def on_current_account_string(self, instance, address):
-        self.show_address(address)
-
     def on_current_account(self, instance, account):
         address = "0x" + account.address.encode("hex")
-        self.current_account_string = address
+        self.address_property = address
 
-    def open_account_list(self):
-        def on_selected_item(instance, value):
-            self.current_account = value.account
-        self.controller.open_account_list_helper(on_selected_item)
+    def on_address_property(self, instance, value):
+        self.show_address(value)
 
 
 class History(BoxLayout):
 
-    current_account = ObjectProperty(None, allownone=True)
+    current_account = ObjectProperty()
+
+    def __init__(self, **kwargs):
+        super(History, self).__init__(**kwargs)
+        Clock.schedule_once(lambda dt: self.setup())
+
+    def setup(self):
+        """
+        Binds Controller.current_account property.
+        """
+        self.controller = App.get_running_app().controller
+        self.controller.bind(current_account=self.setter('current_account'))
+        # triggers the update
+        self.current_account = self.controller.current_account
 
     def on_current_account(self, instance, account):
         self._load_history()
@@ -290,22 +355,68 @@ class History(BoxLayout):
         self.update_history_list(list_items)
 
 
+class SwitchAccount(BoxLayout):
+
+    def on_release(self, list_item):
+        """
+        Sets Controller.current_account and switches to previous screen.
+        """
+        # sets Controller.current_account
+        self.selected_list_item = list_item
+        self.controller.current_account = list_item.account
+        # switches to previous screen
+        self.controller.screen_manager_previous()
+
+    def create_item(self, account):
+        """
+        Creates an account list item from given account.
+        """
+        address = "0x" + account.address.encode("hex")
+        list_item = OneLineListItem(text=address)
+        # makes sure the address doesn't overlap on small screen
+        list_item.ids._lbl_primary.shorten = True
+        list_item.account = account
+        list_item.bind(on_release=lambda x: self.on_release(x))
+        return list_item
+
+    def load_account_list(self):
+        """
+        Fills account list widget from library account list.
+        """
+        self.controller = App.get_running_app().controller
+        account_list_id = self.ids.account_list_id
+        account_list_id.clear_widgets()
+        accounts = self.controller.pywalib.get_account_list()
+        for account in accounts:
+            list_item = self.create_item(account)
+            account_list_id.add_widget(list_item)
+
+
 class Overview(BoxLayout):
 
-    current_account = ObjectProperty(None, allownone=True)
+    current_account = ObjectProperty()
     current_account_string = StringProperty()
-    balance_property = NumericProperty(0)
+
+    def __init__(self, **kwargs):
+        super(Overview, self).__init__(**kwargs)
+        Clock.schedule_once(lambda dt: self.setup())
+
+    def setup(self):
+        """
+        Binds Controller.current_account property.
+        """
+        self.controller = App.get_running_app().controller
+        self.controller.bind(current_account=self.setter('current_account'))
+        # triggers the update
+        self.current_account = self.controller.current_account
 
     def on_current_account(self, instance, account):
+        """
+        Updates current_account_string and fetches the new account balance.
+        """
         address = "0x" + account.address.encode("hex")
         self.current_account_string = address
-
-    def open_account_list(self):
-        controller = App.get_running_app().controller
-        controller.open_account_list_overview()
-
-    def get_title(self):
-        return "%s ETH" % self.balance_property
+        self.controller.fetch_balance()
 
 
 class PWSelectList(BoxLayout):
@@ -351,8 +462,8 @@ class ImportKeystore(BoxLayout):
 # TODO: create a generic account form
 class ManageExisting(BoxLayout):
 
-    current_account = ObjectProperty(None, allownone=True)
-    current_account_string = StringProperty()
+    current_account = ObjectProperty()
+    address_property = StringProperty()
     current_password = StringProperty()
     new_password1 = StringProperty()
     new_password2 = StringProperty()
@@ -363,13 +474,13 @@ class ManageExisting(BoxLayout):
 
     def setup(self):
         """
-        Default state setup.
+        Binds Controller.current_account property.
         """
         self.controller = App.get_running_app().controller
-        try:
-            self.current_account = self.controller.pywalib.get_main_account()
-        except IndexError:
-            pass
+        self.pywalib = self.controller.pywalib
+        self.controller.bind(current_account=self.setter('current_account'))
+        # triggers the update
+        self.current_account = self.controller.current_account
 
     def verify_current_password_field(self):
         """
@@ -387,9 +498,11 @@ class ManageExisting(BoxLayout):
 
     def verify_password_field(self):
         """
-        Makes sure passwords are matching.
+        Makes sure passwords are matching and are not void.
         """
-        return self.new_password1 == self.new_password2
+        passwords_matching = self.new_password1 == self.new_password2
+        passwords_not_void = self.new_password1 != ''
+        return passwords_matching and passwords_not_void
 
     def verify_fields(self):
         """
@@ -397,11 +510,41 @@ class ManageExisting(BoxLayout):
         """
         return self.verify_password_field()
 
-    def delete_account(self):
+    def show_redirect_dialog(self):
+        title = "Account deleted, redirecting..."
+        body = ""
+        body += "Your account was deleted, "
+        body += "you will be redirected to the overview."
+        dialog = Controller.create_dialog(title, body)
+        dialog.open()
+
+    def on_delete_account_yes(self, dialog):
+        """
+        Deletes the account, discarts the warning dialog,
+        shows an info popup and redirects to the landing page.
+        """
+        account = self.current_account
+        self.pywalib.delete_account(account)
+        dialog.dismiss()
+        self.show_redirect_dialog()
+        self.controller.load_landing_page()
+
+    def prompt_delete_account_dialog(self):
         """
         Not yet implemented.
         """
-        Controller.show_not_implemented_dialog()
+        title = "Are you sure?"
+        body = ""
+        body += "This action cannot be undone.\n"
+        body += "Are you sure you want to delete this account?\n"
+        dialog = Controller.create_dialog_helper(title, body)
+        dialog.add_action_button(
+                "No",
+                action=lambda *x: dialog.dismiss())
+        dialog.add_action_button(
+                "Yes",
+                action=lambda *x: self.on_delete_account_yes(dialog))
+        dialog.open()
 
     @run_in_thread
     def update_password(self):
@@ -424,12 +567,7 @@ class ManageExisting(BoxLayout):
 
     def on_current_account(self, instance, account):
         address = "0x" + account.address.encode("hex")
-        self.current_account_string = address
-
-    def open_account_list(self):
-        def on_selected_item(instance, value):
-            self.current_account = value.account
-        self.controller.open_account_list_helper(on_selected_item)
+        self.address_property = address
 
 
 class CreateNewAccount(BoxLayout):
@@ -454,13 +592,15 @@ class CreateNewAccount(BoxLayout):
         self.speed_slider = self.ids.speed_slider_id
         self.security_slider.value = self.speed_slider.value = 50
         self.controller = App.get_running_app().controller
-        # self.toggle_advanced(False)
+        self.toggle_advanced(False)
 
     def verify_password_field(self):
         """
-        Makes sure passwords are matching.
+        Makes sure passwords are matching and are not void.
         """
-        return self.new_password1 == self.new_password2
+        passwords_matching = self.new_password1 == self.new_password2
+        passwords_not_void = self.new_password1 != ''
+        return passwords_matching and passwords_not_void
 
     def verify_fields(self):
         """
@@ -491,14 +631,43 @@ class CreateNewAccount(BoxLayout):
             dialog.open()
             return
 
+    @mainthread
+    def on_account_created(self, account):
+        """
+        Switches to the newly created account.
+        Clears the form.
+        """
+        self.controller.current_account = account
+        self.new_password1 = ''
+        self.new_password2 = ''
+
+    @mainthread
+    def toggle_widgets(self, enabled):
+        """
+        Enables/disables account creation widgets.
+        """
+        self.disabled = not enabled
+
+    def show_redirect_dialog(self):
+        title = "Account created, redirecting..."
+        body = ""
+        body += "Your account was created, "
+        body += "you will be redirected to the overview."
+        dialog = Controller.create_dialog(title, body)
+        dialog.open()
+
     @run_in_thread
     def create_account(self):
         """
         Creates an account from provided form.
         Verify we can unlock it.
+        Disables widgets during the process, so the user doesn't try
+        to create another account during the process.
         """
+        self.toggle_widgets(False)
         if not self.verify_fields():
             Controller.show_invalid_form_dialog()
+            self.toggle_widgets(True)
             return
         pywalib = self.controller.pywalib
         password = self.new_password1
@@ -510,7 +679,11 @@ class CreateNewAccount(BoxLayout):
         account = pywalib.new_account(
                 password=password, security_ratio=security_ratio)
         Controller.snackbar_message("Created!")
+        self.toggle_widgets(True)
+        self.on_account_created(account)
         CreateNewAccount.try_unlock(account, password)
+        self.show_redirect_dialog()
+        self.controller.load_landing_page()
         return account
 
     def toggle_advanced(self, show):
@@ -519,13 +692,10 @@ class CreateNewAccount(BoxLayout):
         https://stackoverflow.com/q/23211142/185510
         """
         advanced = self.ids.advanced_id
-        if not show:
-            # save the old y-coordinate
-            advanced.saved_y = advanced.y
-            # now move the widget offscreen
-            advanced.y = 5000
-        else:
-            advanced.y = advanced.saved_y
+        alpha = 1 if show else 0
+        for widget in advanced.children:
+            widget.canvas.opacity = alpha
+            widget.disabled = not show
 
 
 class AddressButton(MDFlatButton):
@@ -552,6 +722,9 @@ class AddressButton(MDFlatButton):
             # parent_width and actual content size
             content.width = parent_width - button_margin
         self.parent.bind(size=on_parent_size)
+        # call it once manually, refs:
+        # https://github.com/AndreMiras/PyWallet/issues/74
+        on_parent_size(self.parent, None)
 
 
 class PWToolbar(Toolbar):
@@ -565,17 +738,7 @@ class PWToolbar(Toolbar):
     def setup(self):
         self.controller = App.get_running_app().controller
         self.navigation = self.controller.ids.navigation_id
-        self.screen_manager = self.controller.ids.screen_manager_id
-        # bind balance update to title
-        overview = self.controller.overview
-        overview.bind(
-            balance_property=lambda *x: self.on_overview_balance_property())
-        # let's add the default title while waiting for the update
-        self.title_property = self.controller.get_overview_title()
         self.load_default_navigation()
-
-    def on_overview_balance_property(self):
-        self.title_property = self.controller.get_overview_title()
 
     def load_default_navigation(self):
         self.left_action_items = [
@@ -648,6 +811,7 @@ class AboutOverview(BoxLayout):
 
     def load_about(self):
         self.about_text_property = "" + \
+            "PyWallet version: %s\n" % (__version__) + \
             "Project source code and info available on GitHub at: \n" + \
             "[color=00BFFF][ref=github]" + \
             self.project_page_property + \
@@ -671,7 +835,6 @@ class AboutDiagnostic(BoxLayout):
         """
         Controller.patch_keystore_path()
         test_suite = suite()
-        print("test_suite:", test_suite)
         self.stream_property = ""
         stream = StringIOCBWrite(callback_write=self.callback_write)
         verbosity = 2
@@ -679,9 +842,30 @@ class AboutDiagnostic(BoxLayout):
                 stream=stream, verbosity=verbosity).run(test_suite)
 
 
+class OverviewScreen(Screen):
+
+    title_property = StringProperty()
+
+    def set_title(self, title):
+        self.title_property = title
+
+
+class SwitchAccountScreen(Screen):
+    pass
+
+
+class ManageKeystoreScreen(Screen):
+    pass
+
+
+class AboutScreen(Screen):
+    pass
+
+
 class Controller(FloatLayout):
 
-    current_account = ObjectProperty(None, allownone=True)
+    current_account = ObjectProperty()
+    current_account_balance = NumericProperty(0)
     # keeps track of all dialogs alive
     dialogs = []
 
@@ -689,62 +873,96 @@ class Controller(FloatLayout):
         super(Controller, self).__init__(**kwargs)
         keystore_path = Controller.get_keystore_path()
         self.pywalib = PyWalib(keystore_path)
+        self.screen_history = []
         Clock.schedule_once(lambda dt: self.load_landing_page())
 
     @property
     def overview(self):
-        overview_bnavigation_id = self.ids.overview_bnavigation_id
-        return overview_bnavigation_id.ids.overview_id
+        overview_screen = self.ids.overview_screen_id
+        overview_bnavigation = overview_screen.ids.overview_bnavigation_id
+        return overview_bnavigation.ids.overview_id
 
     @property
     def history(self):
         return self.overview.ids.history_id
 
     @property
+    def switch_account(self):
+        screen_manager = self.screen_manager
+        switch_account_screen = screen_manager.get_screen('switch_account')
+        switch_account_id = switch_account_screen.ids.switch_account_id
+        return switch_account_id
+
+    @property
     def send(self):
-        overview_bnavigation_id = self.ids.overview_bnavigation_id
+        screen_manager = self.screen_manager
+        overview_screen = screen_manager.get_screen('overview')
+        overview_bnavigation_id = overview_screen.ids.overview_bnavigation_id
         return overview_bnavigation_id.ids.send_id
+
+    @property
+    def manage_keystores(self):
+        screen_manager = self.screen_manager
+        manage_keystores_screen = screen_manager.get_screen('manage_keystores')
+        manage_keystores_bnavigation_id = \
+            manage_keystores_screen.ids.manage_keystores_id
+        return manage_keystores_bnavigation_id
+
+    @property
+    def manage_existing(self):
+        manage_keystores = self.manage_keystores
+        return manage_keystores.ids.manage_existing_id
+
+    @property
+    def create_new_account(self):
+        manage_keystores = self.manage_keystores
+        return manage_keystores.ids.create_new_account_id
 
     @property
     def toolbar(self):
         return self.ids.toolbar_id
 
+    @property
+    def screen_manager(self):
+        return self.ids.screen_manager_id
+
     def set_toolbar_title(self, title):
         self.toolbar.title_property = title
 
-    def open_account_list_helper(self, on_selected_item):
-        title = "Select account"
-        items = []
-        pywalib = self.pywalib
-        account_list = pywalib.get_account_list()
-        for account in account_list:
-            address = '0x' + account.address.encode("hex")
-            item = OneLineListItem(text=address)
-            # makes sure the address doesn't wrap in multiple lines,
-            # but gets shortened
-            item.ids._lbl_primary.shorten = True
-            item.account = account
-            items.append(item)
-        dialog = Controller.create_list_dialog(
-            title, items, on_selected_item)
-        dialog.open()
-
-    def open_account_list_overview(self):
-        def on_selected_item(instance, value):
-            self.set_current_account(value.account)
-        self.open_account_list_helper(on_selected_item)
-
-    def set_current_account(self, account):
-        self.current_account = account
-
-    def on_current_account(self, instance, value):
+    def bind_current_account_balance(self):
         """
-        Updates Overview.current_account and History.current_account,
-        then fetch account data.
+        Binds the current_account_balance to the Toolbar title.
         """
-        self.overview.current_account = value
-        self.history.current_account = value
-        self._load_balance()
+        self.bind(current_account_balance=self.update_toolbar_title_balance)
+
+    def unbind_current_account_balance(self):
+        """
+        Unbinds the current_account_balance from the Toolbar title.
+        """
+        self.unbind(current_account_balance=self.update_toolbar_title_balance)
+
+    def screen_manager_current(self, current, direction=None):
+        screens = {
+            'overview': OverviewScreen,
+            'switch_account': SwitchAccountScreen,
+            'manage_keystores': ManageKeystoreScreen,
+            'about': AboutScreen,
+        }
+        screen_manager = self.screen_manager
+        if not screen_manager.has_screen(current):
+            screen = screens[current](name=current)
+            screen_manager.add_widget(screen)
+        if direction is not None:
+            screen_manager.transition.direction = direction
+        screen_manager.current = current
+        self.screen_history.append(current)
+
+    def screen_manager_previous(self):
+        try:
+            previous_screen = self.screen_history[-2]
+        except IndexError:
+            previous_screen = 'overview'
+        self.screen_manager_current(previous_screen, direction='right')
 
     @staticmethod
     def show_invalid_form_dialog():
@@ -781,30 +999,6 @@ class Controller(FloatLayout):
         return os.path.dirname(os.path.abspath(__file__))
 
     @staticmethod
-    def create_list_dialog(title, items, on_selected_item):
-        """
-        Creates a dialog from given title and list.
-        items is a list of BaseListItem objects.
-        """
-        # select_list = PWSelectList(items=items, on_release=on_release)
-        select_list = PWSelectList(items=items)
-        select_list.bind(selected_item=on_selected_item)
-        content = select_list
-        dialog = MDDialog(
-                        title=title,
-                        content=content,
-                        size_hint=(.9, .9))
-        # workaround for MDDialog container size (too small by default)
-        dialog.ids.container.size_hint_y = 1
-        # close the dialog as we select the element
-        select_list.bind(
-            selected_item=lambda instance, value: dialog.dismiss())
-        dialog.add_action_button(
-                "Dismiss",
-                action=lambda *x: dialog.dismiss())
-        return dialog
-
-    @staticmethod
     def on_dialog_dismiss(dialog):
         """
         Removes it from the dialogs track list.
@@ -821,7 +1015,7 @@ class Controller(FloatLayout):
             dialog.dispatch('on_dismiss')
 
     @staticmethod
-    def create_dialog(title, body):
+    def create_dialog_helper(title, body):
         """
         Creates a dialog from given title and body.
         Adds it to the dialogs track list.
@@ -837,13 +1031,23 @@ class Controller(FloatLayout):
                         title=title,
                         content=content,
                         size_hint=(.8, None),
-                        height=dp(200),
+                        height=dp(250),
                         auto_dismiss=False)
+        dialog.bind(on_dismiss=Controller.on_dialog_dismiss)
+        Controller.dialogs.append(dialog)
+        return dialog
+
+    @staticmethod
+    def create_dialog(title, body):
+        """
+        Creates a dialog from given title and body.
+        Adds it to the dialogs track list.
+        Appends dismiss action.
+        """
+        dialog = Controller.create_dialog_helper(title, body)
         dialog.add_action_button(
                 "Dismiss",
                 action=lambda *x: dialog.dismiss())
-        dialog.bind(on_dismiss=Controller.on_dialog_dismiss)
-        Controller.dialogs.append(dialog)
         return dialog
 
     @staticmethod
@@ -868,13 +1072,9 @@ class Controller(FloatLayout):
         dialog.open()
 
     @mainthread
-    def update_balance_label(self, balance):
-        overview_id = self.overview
-        overview_id.balance_property = balance
-
-    def get_overview_title(self):
-        overview_id = self.overview
-        return overview_id.get_title()
+    def update_toolbar_title_balance(self, instance=None, value=None):
+        title = "%s ETH" % (self.current_account_balance)
+        self.set_toolbar_title(title)
 
     @staticmethod
     @mainthread
@@ -888,28 +1088,35 @@ class Controller(FloatLayout):
         try:
             # will trigger account data fetching
             self.current_account = self.pywalib.get_main_account()
-            self.ids.screen_manager_id.current = "overview"
-            self.ids.screen_manager_id.transition.direction = "right"
+            self.screen_manager_current('overview')
         except IndexError:
             self.load_create_new_account()
 
-    @run_in_thread
-    def _load_balance(self):
+    def fetch_balance(self):
+        """
+        Fetches the new balance and current_account_balance property.
+        """
         account = self.current_account
         try:
-            balance = self.pywalib.get_balance(account.address.encode("hex"))
+            self.current_account_balance = self.pywalib.get_balance(
+                account.address.encode("hex"))
         except ConnectionError:
             Controller.on_balance_connection_error()
             return
-        self.update_balance_label(balance)
+
+    def load_switch_account(self):
+        """
+        Loads the switch account screen.
+        """
+        # loads the switch account screen
+        self.screen_manager_current('switch_account', direction='left')
 
     def load_manage_keystores(self):
         """
         Loads the manage keystores screen.
         """
         # loads the manage keystores screen
-        self.ids.screen_manager_id.transition.direction = "left"
-        self.ids.screen_manager_id.current = 'manage_keystores'
+        self.screen_manager_current('manage_keystores', direction='left')
 
     def load_create_new_account(self):
         """
@@ -917,16 +1124,26 @@ class Controller(FloatLayout):
         """
         self.load_manage_keystores()
         # loads the create new account tab
-        manage_keystores = self.ids.manage_keystores_id
-        create_new_account = manage_keystores.ids.create_new_account_id
-        create_new_account.dispatch('on_tab_press')
+        manage_keystores = self.manage_keystores
+        create_new_account_nav_item = \
+            manage_keystores.ids.create_new_account_nav_item_id
+        create_new_account_nav_item.dispatch('on_tab_press')
 
     def load_about_screen(self):
         """
         Loads the about screen.
         """
-        self.ids.screen_manager_id.transition.direction = "left"
-        self.ids.screen_manager_id.current = "about"
+        self.screen_manager_current('about', direction='left')
+
+
+class DebugRavenClient(object):
+    """
+    The DebugRavenClient should be used in debug mode, it just raises
+    the exception rather than capturing it.
+    """
+
+    def captureException(self):
+        raise
 
 
 class PyWalletApp(App):
@@ -941,7 +1158,10 @@ class PyWalletApp(App):
         return self.root
 
 
-def configure_sentry():
+def configure_sentry(in_debug=False):
+    """
+    Configure the Raven client, or create a dummy one if `in_debug` is `True`.
+    """
     key = 'eaee971c463b49678f6f352dfec497a9'
     # the public DSN URL is not available on the Python client
     # so we're exposing the secret and will be revoking it on abuse
@@ -950,13 +1170,23 @@ def configure_sentry():
     project_id = '191660'
     dsn = 'https://{key}:{secret}@sentry.io/{project_id}'.format(
         key=key, secret=secret, project_id=project_id)
-    client = Client(dsn)
+    if in_debug:
+        client = DebugRavenClient()
+    else:
+        client = Client(dsn)
     return client
 
 
 if __name__ == '__main__':
-    client = configure_sentry()
+    # when the -d/--debug flag is set, Kivy sets log level to debug
+    level = Logger.getEffectiveLevel()
+    in_debug = level == LOG_LEVELS.get('debug')
+    client = configure_sentry(in_debug)
     try:
         PyWalletApp().run()
     except:
+        if type(client) == Client:
+            Logger.info(
+                'Errors will be sent to Sentry, run with "--debug" if you '
+                'are a developper and want to the error in the shell.')
         client.captureException()